app/pyproject.toml
[project]
name = "nicegui-component-based"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nicegui",
    "nicegui[highcharts]",
    "pyinstaller>=6.13.0",
    "pywebview>=5.4",
    "requests>=2.31.0",
    "authlib>=1.2.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]
